setup and remove mostly working now

+++ setup.sh
@@ -11,12 +11,31 @@ RED_UL='\033[38;4;31m'
 YELLOW='\033[38;33m'
 YELLOW_UL='\033[38;4;33m'
 
+# Ensure script is not being run with root privileges
+if [ $EUID -eq 0 ]; then
+  echo "Please don't run this script with root privileges!"
+  exit 1
+fi
+
+SUDO_ON=$(sudo -n command &>/dev/null; echo $?)
+if [ $SUDO_ON -gt 0 ]; then
+  echo "Some of the operations must be run as admin so please enter your admin password: "
+  sudo -v
+fi
+
+# Keep-alive: update existing `sudo` time stamp until this script has finished
+while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
+
+# Close any open System Preferences panes, to prevent them from overriding settings we’re about to change
+echo "Closing any open System Preferences dialogues"
+osascript -e 'tell application "System Preferences" to quit'
+
 # Give the computer a name
 read -p "Pick a name for this machine [$(hostname)]:  " COMPUTER_NAME
 if [ -z "$COMPUTER_NAME" ]; then
   COMPUTER_NAME=$(hostname)
 fi
-echo -e "Using "$BLUE"$COMPUTER_NAME"$ESC" for this machine"
+echo -e "Using "$GREEN"$COMPUTER_NAME"$ESC" for this machine"
 
 nixConfig="./hosts/$COMPUTER_NAME/configuration.nix"
 if [ ! -f "./hosts/$COMPUTER_NAME/configuration.nix" ]; then
@@ -24,19 +43,10 @@ if [ ! -f "./hosts/$COMPUTER_NAME/configuration.nix" ]; then
   echo "{}" > "$nixConfig"
 fi
 
-# Ask for the administrator password upfront
-echo "Some of the operations must be run as admin so please enter your admin password."
-sudo -v
-
-# Keep-alive: update existing `sudo` time stamp until `.macos` has finished
-while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
-
-# Close any open System Preferences panes, to prevent them from overriding settings we’re about to change
-osascript -e 'tell application "System Preferences" to quit'
-
-# the version of curl that comes with macosx is ancient cannot deal with IPV6 addresses so need disable them for now
-networksetup -setv6off Wi-Fi &>/dev/null
-networksetup -setv6off Ethernet &>/dev/null
+# the version of curl that comes with macosx is ancient cannot deal with IPV6 addresses
+# so need disable them for now. Seems be an issue on some corporate networks for some reason.
+sudo networksetup -setv6off Wi-Fi &>/dev/null
+sudo networksetup -setv6off Ethernet &>/dev/null
 
 # Set computer name (as done via System Preferences → Sharing)
 sudo scutil --set ComputerName "$COMPUTER_NAME"
@@ -49,55 +59,76 @@ dscacheutil -flushcache
 sudo nvram SystemAudioVolume=" "
 
 # Nix
-#curl https://nixos.org/nix/install | sh
+if [[ ! $(type nix-env 2>/dev/null) ]]; then
+  curl https://nixos.org/nix/install | sh
 
-# Create this empty directory to prevent warnings coming from nix
-mkdir -p /nix/var/nix/profiles/per-user/root/channels
-sudo ln -s private/var/run /run
+  if [ ! -e private/var/run ]; then
+    sudo ln -s private/var/run /run
+  fi
 
-NIX_INIT_SCRIPT="$HOME/.nix-profile/etc/profile.d/nix.sh"
-if [ -f $NIX_INIT_SCRIPT ]; then
-  source /Users/simon/.nix-profile/etc/profile.d/nix.sh
-else
-  echo -e ""$RED"Nix doesn't appear to have been installed correctly. Aborting!"$ESC""
-  echo "$NIX_INIT_SCRIPT should exist and be executable - please try again."
-  exit
+  # Create this empty directory to prevent warnings coming from nix
+  mkdir -p /nix/var/nix/profiles/per-user/root/channels
+
+  NIX_INIT_SCRIPT="$HOME/.nix-profile/etc/profile.d/nix.sh"
+  if [ -f $NIX_INIT_SCRIPT ]; then
+    source "$NIX_INIT_SCRIPT"
+  else
+    echo -e ""$RED"Nix doesn't appear to have been installed correctly. Aborting!"$ESC""
+    echo "$NIX_INIT_SCRIPT should exist and be executable - please try again."
+    exit 1
+  fi
+
+  # Ensure Nix has already been installed
+  if [[ ! $(type nix-env 2>/dev/null) ]]; then
+    echo -e "Cannot find "$YELLOW"nix-env"$ESC" in the PATH"
+    echo "This means that the nix init script has not been sourced properly"
+    exit 1
+  fi
 fi
+
 
 # sets the host for the nix/hostname.nix script to pick up instead of asking the system for it
 # this is because the host may not be fully setup until after a reboot occurs
 export HOST="$COMPUTER_NAME"
 
-if [ -f /etc/shells ]; then
+if [ ! -L /etc/shells -a -f /etc/shells ]; then
   sudo mv /etc/shells /etc/shells.bak
 fi
-if [ -f /etc/zprofile ]; then
+if [ ! -L /etc/zprofile -a -f /etc/zprofile ]; then
   sudo mv /etc/zprofile /etc/zprofile.local
 fi
-if [ -f /etc/zshrc ]; then
+if [ ! -L /etc/zshrc -a -f /etc/zshrc ]; then
   sudo mv /etc/zshrc /etc/zshrc.local
 fi
-
-# setup the nix-daemon users - without these nix-darwin will refuse to complete its setup steps
-curl -L -o bootstrap.sh https://raw.githubusercontent.com/LnL7/nix-darwin/master/bootstrap.sh
-chmod +x bootstrap.sh
-./bootstrap.sh -u
-rm bootstrap.sh
 
 # home-manager
 nix-channel --add https://github.com/rycee/home-manager/archive/release-19.09.tar.gz home-manager
 nix-channel --update
 
-if [ -L /run/current-system/sw/bin/bash ]; then
-  echo "Switching default shell to newer nix supplied bash"
-  chsh -s /run/current-system/sw/bin/bash
+# nix darwin
+if [[ ! $(type darwin-rebuild 2>/dev/null) ]]; then
+  nix-build https://github.com/LnL7/nix-darwin/archive/master.tar.gz -A installer
+  TERM=xterm ./result/bin/darwin-installer
+
+  # Ensure nix-darwin has already been installed
+  if [[ ! $(type darwin-rebuild 2>/dev/null) ]]; then
+    echo -e "Cannot find "$YELLOW"darwin-rebuild"$ESC" in the PATH"
+    echo "This means that the nix-darwin init script has not been sourced properly"
+    exit 1
+  fi
 fi
+
+NIX_SUPPLIED_BASH="/run/current-system/sw/bin/bash"
+if [[ "$SHELL" != "/run"* && -L "$NIX_SUPPLIED_BASH" ]]; then
+  echo "Switching default shell to newer nix supplied bash"
+  chsh -s "$NIX_SUPPLIED_BASH"
+  export SHELL="$NIX_SUPPLIED_BASH"
+fi
+
+darwin-rebuild switch
+echo -e ""$GREEN"Successfully completed!"$ESC""
+echo "Restart your machine for GPG/keyring to be setup properly"
 
 # re-enable IPV6 now we have a decent curl from nixpkgs
-networksetup -setv6automatic Wi-Fi &>/dev/null
-networksetup -setv6automatic Ethernet &>/dev/null
-
-if [[ $(type darwin-rebuild 2>/dev/null) ]]; then
-  echo -e ""$GREEN"Successfully completed!"$ESC""
-  echo "Restart your machine for GPG/keyring to be setup properly"
-fi
+sudo networksetup -setv6automatic Wi-Fi &>/dev/null
+sudo networksetup -setv6automatic Ethernet &>/dev/null
